패키지 판매 리스트 및 하이퍼링크 수정

+++ src/main/java/com/project/tour/repository/AdminPackageDateRepository.java
@@ -6,8 +6,9 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdminPackageDateRepository extends JpaRepository<PackageDate, Long> {
+public interface AdminPackageDateRepository extends JpaRepository<PackageDate, Integer> {
     Page<PackageDate> findAll(Pageable pageable);
     Page<PackageDate> findByPackages_Id(Long id,Pageable pageable);
+
 
 }
